INS-213: Move public interface tests on storage level
